update compatible PostgreSql

+++ packages/_leaa-common/src/entrys/article.entity.ts
@@ -33,7 +33,7 @@ export class Article extends Base {
   @Field(() => String, { nullable: true })
   public content?: string;
 
-  @Column({ type: 'tinyint', default: 0 })
+  @Column({ type: 'int', default: 0 })
   @Field(() => Int, { nullable: true })
   public status?: number;
